chore: refactor PasswordMeter as a child component
