+++ lib/components/PasswordInput/PasswordConstants.ts
@@ -1,3 +1,5 @@
+export type TScore = 0 | 1 | 2 | 3 | 4;
+
 export type passwordKeys =
   | "common"
   | "commonNames"
